+++ creme/linear_model/log_reg.py
@@ -13,7 +13,7 @@ class LogisticRegression(base.BinaryClassifier):
 
     Parameters:
         optimizer (optim.Optimizer): The sequential optimizer used to find the best weights.
-        loss (optim.Loss): The loss function to optimize for.
+        loss (optim.BinaryClassificationLoss): The loss function to optimize for.
         l2 (float): regularization amount used to push weights towards 0.
 
     Attributes:
@@ -53,28 +53,26 @@ class LogisticRegression(base.BinaryClassifier):
         self.l2 = l2
         self.weights = collections.defaultdict(float)
 
-    def _predict_proba_with_weights(self, x, w):
-        return utils.sigmoid(utils.dot(x, w))
-
-    def _calc_gradient(self, y_true, y_pred, loss, x, w):
-        """Returns a dictionary containing the gradient w.r.t. each feature."""
-        loss_gradient = loss.gradient(y_true=y_true, y_pred=y_pred)
-        return {i: xi * loss_gradient + self.l2 * w.get(i, 0) for i, xi in x.items()}
-
     def fit_one(self, x, y):
 
-        # Update the weights with the error gradient
-        self.weights, y_pred = self.optimizer.update_weights(
-            x=x,
-            y=y,
-            w=self.weights,
-            loss=self.loss,
-            f_pred=self._predict_proba_with_weights,
-            f_grad=self._calc_gradient
-        )
+        # Some optimizers need to do something before a prediction is made
+        self.weights = self.optimizer.update_before_pred(w=self.weights)
+
+        # Make a prediction for the given features
+        y_pred = self.predict_proba_one(x)[True]
+
+        # Compute the gradient w.r.t. each feature
+        loss_gradient = self.loss.gradient(y_true=y, y_pred=y_pred)
+        gradient = {
+            i: xi * loss_gradient + self.l2 * self.weights.get(i, 0)
+            for i, xi in x.items()
+        }
+
+        # Update the weights by using the gradient
+        self.weights = self.optimizer.update_after_pred(g=gradient, w=self.weights)
 
         return self
 
     def predict_proba_one(self, x):
-        y_pred = self._predict_proba_with_weights(x, self.weights)
+        y_pred = utils.sigmoid(utils.dot(x, self.weights))
         return {False: 1. - y_pred, True: y_pred}