+++ creme/optim/ada_delta.py
@@ -53,7 +53,7 @@ class AdaDelta(base.Optimizer):
     def _rms(self, x):
         return (x + self.eps) ** 0.5
 
-    def update_weights_with_gradient(self, w, g):
+    def _update_after_pred(self, w, g):
 
         for i, gi in g.items():