+++ creme/optim/ada_grad.py
@@ -44,7 +44,7 @@ class AdaGrad(base.Optimizer):
         self.eps = eps
         self.g2 = collections.defaultdict(float)
 
-    def update_weights_with_gradient(self, w, g):
+    def _update_after_pred(self, w, g):
 
         for i, gi in g.items():
             self.g2[i] += gi ** 2