+++ creme/reco/sgd_baseline.py
@@ -70,6 +70,9 @@ class SGDBaseline(base.Recommender):
 
     def fit_one(self, r_id, c_id, y):
 
+        self.row_optimizer.update_before_pred(self.row_biases)
+        self.col_optimizer.update_before_pred(self.col_biases)
+
         # Predict the value
         y_pred = self.predict_one(r_id, c_id)
 
@@ -78,11 +81,11 @@ class SGDBaseline(base.Recommender):
 
         # Update the row biases
         r_grad = {r_id: loss_gradient + self.l2 * self.row_biases.get(r_id, 0)}
-        self.row_biases = self.row_optimizer.update_weights_with_gradient(self.row_biases, r_grad)
+        self.row_biases = self.row_optimizer.update_after_pred(self.row_biases, r_grad)
 
         # Update the row biases
         c_grad = {c_id: loss_gradient + self.l2 * self.col_biases.get(c_id, 0)}
-        self.col_biases = self.col_optimizer.update_weights_with_gradient(self.col_biases, c_grad)
+        self.col_biases = self.col_optimizer.update_after_pred(self.col_biases, c_grad)
 
         # Update the global mean
         self.global_mean.update(y)